........
  r1039 | weil | 2008-02-01 22:54:16 +0100 (Fr, 01 Feb 2008) | 1 line
  
  cvs update
........
  r1040 | weil | 2008-02-01 22:56:29 +0100 (Fr, 01 Feb 2008) | 1 line
  
  cvs update (tcg)
........
  r1041 | weil | 2008-02-02 08:47:10 +0100 (Sa, 02 Feb 2008) | 1 line
  
  cvs update
........
  r1042 | weil | 2008-02-03 11:31:48 +0100 (So, 03 Feb 2008) | 1 line
  
  cvs update
........
  r1043 | weil | 2008-02-03 11:37:28 +0100 (So, 03 Feb 2008) | 1 line
  
  cvs update
........
  r1044 | weil | 2008-02-03 11:38:23 +0100 (So, 03 Feb 2008) | 1 line
  
  fixed spelling
........
  r1046 | weil | 2008-02-04 00:26:38 +0100 (Mo, 04 Feb 2008) | 1 line
  
  cvs update
........
  r1047 | weil | 2008-02-04 19:10:39 +0100 (Mo, 04 Feb 2008) | 1 line
  
  cvs update
........
  r1048 | weil | 2008-02-09 07:04:14 +0100 (Sa, 09 Feb 2008) | 1 line
  
  cvs update
........
  r1049 | weil | 2008-02-10 20:31:49 +0100 (So, 10 Feb 2008) | 1 line
  
  cvs update
........
  r1050 | weil | 2008-02-12 20:02:51 +0100 (Di, 12 Feb 2008) | 1 line
  
  cvs update
........
  r1051 | weil | 2008-02-13 19:36:45 +0100 (Mi, 13 Feb 2008) | 1 line
  
  cvs update
........
  r1052 | weil | 2008-02-14 21:06:40 +0100 (Do, 14 Feb 2008) | 1 line
  
  cvs update
........
  r1053 | weil | 2008-02-17 10:47:16 +0100 (So, 17 Feb 2008) | 1 line
  
  cvs update
........
  r1054 | weil | 2008-02-17 13:09:31 +0100 (So, 17 Feb 2008) | 1 line
  
  cvs update
........
  r1055 | weil | 2008-02-24 20:41:35 +0100 (So, 24 Feb 2008) | 1 line
  
  cvs update
........
  r1056 | weil | 2008-02-28 19:10:35 +0100 (Do, 28 Feb 2008) | 1 line
  
  cvs update
........
  r1061 | weil | 2008-02-29 21:28:29 +0100 (Fr, 29 Feb 2008) | 1 line
  
  cvs update
........
  r1062 | weil | 2008-03-01 20:50:23 +0100 (Sa, 01 Mär 2008) | 1 line
  
  cvs update
........
  r1064 | weil | 2008-03-04 22:38:38 +0100 (Di, 04 Mär 2008) | 1 line
  
  cvs update
........
  r1065 | weil | 2008-03-04 22:39:22 +0100 (Di, 04 Mär 2008) | 1 line
  
  cvs update
........
  r1067 | weil | 2008-03-05 19:15:13 +0100 (Mi, 05 Mär 2008) | 1 line
  
  cvs update
........
  r1068 | weil | 2008-03-06 22:12:12 +0100 (Do, 06 Mär 2008) | 1 line
  
  cvs update
........
  r1069 | weil | 2008-03-06 22:13:35 +0100 (Do, 06 Mär 2008) | 1 line
  
  cvs update
........
  r1070 | weil | 2008-03-08 22:05:28 +0100 (Sa, 08 Mär 2008) | 1 line
  
  cvs update
........
  r1071 | weil | 2008-03-09 17:34:42 +0100 (So, 09 Mär 2008) | 1 line
  
  cvs update
........
  r1072 | weil | 2008-03-10 19:17:19 +0100 (Mo, 10 Mär 2008) | 1 line
  
  cvs update
........
  r1073 | weil | 2008-03-10 21:59:27 +0100 (Mo, 10 Mär 2008) | 1 line
  
  cvs update
........
  r1074 | weil | 2008-03-11 19:26:15 +0100 (Di, 11 Mär 2008) | 1 line
  
  cvs update
........
  r1075 | weil | 2008-03-11 22:15:33 +0100 (Di, 11 Mär 2008) | 1 line
  
  cvs update
........
  r1076 | weil | 2008-03-13 22:24:53 +0100 (Do, 13 Mär 2008) | 1 line
  
  cvs update
........
  r1077 | weil | 2008-03-14 08:11:14 +0100 (Fr, 14 Mär 2008) | 1 line
  
  cvs update
........
  r1078 | weil | 2008-03-15 11:22:36 +0100 (Sa, 15 Mär 2008) | 1 line
  
  cvs update
........
  r1080 | weil | 2008-03-15 22:17:46 +0100 (Sa, 15 Mär 2008) | 1 line
  
  cvs update
........
  r1083 | weil | 2008-03-16 23:12:14 +0100 (So, 16 Mär 2008) | 1 line
  
  cvs update
........
  r1093 | weil | 2008-03-21 21:27:45 +0100 (Fr, 21 Mär 2008) | 1 line
  
  cvs update
........
  r1094 | weil | 2008-03-22 17:18:26 +0100 (Sa, 22 Mär 2008) | 1 line
  
  cvs update
........
  r1095 | weil | 2008-03-24 08:36:38 +0100 (Mo, 24 Mär 2008) | 1 line
  
  cvs update
........
  r1096 | weil | 2008-03-29 21:34:07 +0100 (Sa, 29 Mär 2008) | 1 line
  
  cvs update
........
  r1097 | weil | 2008-03-29 21:35:08 +0100 (Sa, 29 Mär 2008) | 1 line
  
  cvs update
........
  r1098 | weil | 2008-03-30 22:30:32 +0200 (So, 30 Mär 2008) | 1 line
  
  cvs update
........


git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/ar7-firmware/qemu/trunk@1102 11fd3d49-aaf6-0310-8052-ea9566248e5e

+++ vnc.c
@@ -879,8 +879,8 @@ static void pointer_event(VncState *vs, int button_mask, int x, int y)
 	dz = 1;
 
     if (vs->absolute) {
-	kbd_mouse_event(x * 0x7FFF / vs->ds->width,
-			y * 0x7FFF / vs->ds->height,
+	kbd_mouse_event(x * 0x7FFF / (vs->ds->width - 1),
+			y * 0x7FFF / (vs->ds->height - 1),
 			dz, buttons);
     } else if (vs->has_pointer_type_change) {
 	x -= 0x7FFF;
@@ -945,6 +945,7 @@ static void do_key_event(VncState *vs, int down, uint32_t sym)
             return;
         }
         break;
+    case 0x3a:			/* CapsLock */
     case 0x45:			/* NumLock */
         if (!down)
             vs->modifiers_state[keycode] ^= 1;
@@ -1898,6 +1899,22 @@ static int protocol_version(VncState *vs, uint8_t *version, size_t len)
     return 0;
 }
 
+static void vnc_connect(VncState *vs)
+{
+    VNC_DEBUG("New client on socket %d\n", vs->csock);
+    socket_set_nonblock(vs->csock);
+    qemu_set_fd_handler2(vs->csock, NULL, vnc_client_read, NULL, vs);
+    vnc_write(vs, "RFB 003.008\n", 12);
+    vnc_flush(vs);
+    vnc_read_when(vs, protocol_version, 12);
+    memset(vs->old_data, 0, vs->ds->linesize * vs->ds->height);
+    memset(vs->dirty_row, 0xFF, sizeof(vs->dirty_row));
+    vs->has_resize = 0;
+    vs->has_hextile = 0;
+    vs->ds->dpy_copy = NULL;
+    vnc_update_client(vs);
+}
+
 static void vnc_listen_read(void *opaque)
 {
     VncState *vs = opaque;
@@ -1909,18 +1926,7 @@ static void vnc_listen_read(void *opaque)
 
     vs->csock = accept(vs->lsock, (struct sockaddr *)&addr, &addrlen);
     if (vs->csock != -1) {
-	VNC_DEBUG("New client on socket %d\n", vs->csock);
-        socket_set_nonblock(vs->csock);
-	qemu_set_fd_handler2(vs->csock, NULL, vnc_client_read, NULL, opaque);
-	vnc_write(vs, "RFB 003.008\n", 12);
-	vnc_flush(vs);
-	vnc_read_when(vs, protocol_version, 12);
-	memset(vs->old_data, 0, vs->ds->linesize * vs->ds->height);
-	memset(vs->dirty_row, 0xFF, sizeof(vs->dirty_row));
-	vs->has_resize = 0;
-	vs->has_hextile = 0;
-	vs->ds->dpy_copy = NULL;
-        vnc_update_client(vs);
+        vnc_connect(vs);
     }
 }
 
@@ -2087,6 +2093,7 @@ int vnc_display_open(DisplayState *ds, const char *display)
     VncState *vs = ds ? (VncState *)ds->opaque : vnc_state;
     const char *options;
     int password = 0;
+    int reverse = 0;
 #if CONFIG_VNC_TLS
     int tls = 0, x509 = 0;
 #endif
@@ -2103,6 +2110,8 @@ int vnc_display_open(DisplayState *ds, const char *display)
 	options++;
 	if (strncmp(options, "password", 8) == 0) {
 	    password = 1; /* Require password auth */
+	} else if (strncmp(options, "reverse", 7) == 0) {
+	    reverse = 1;
 #if CONFIG_VNC_TLS
 	} else if (strncmp(options, "tls", 3) == 0) {
 	    tls = 1; /* Require TLS */
@@ -2196,7 +2205,9 @@ int vnc_display_open(DisplayState *ds, const char *display)
 	memset(uaddr.sun_path, 0, 108);
 	snprintf(uaddr.sun_path, 108, "%s", p);
 
-	unlink(uaddr.sun_path);
+	if (!reverse) {
+	    unlink(uaddr.sun_path);
+	}
     } else
 #endif
     {
@@ -2210,7 +2221,7 @@ int vnc_display_open(DisplayState *ds, const char *display)
 	    return -1;
 	}
 
-	iaddr.sin_port = htons(ntohs(iaddr.sin_port) + 5900);
+	iaddr.sin_port = htons(ntohs(iaddr.sin_port) + (reverse ? 0 : 5900));
 
 	vs->lsock = socket(PF_INET, SOCK_STREAM, 0);
 	if (vs->lsock == -1) {
@@ -2231,6 +2242,22 @@ int vnc_display_open(DisplayState *ds, const char *display)
 	    vs->display = NULL;
 	    return -1;
 	}
+    }
+
+    if (reverse) {
+        if (connect(vs->lsock, addr, addrlen) == -1) {
+            fprintf(stderr, "Connection to VNC client failed\n");
+            close(vs->lsock);
+            vs->lsock = -1;
+            free(vs->display);
+            vs->display = NULL;
+            return -1;
+        } else {
+            vs->csock = vs->lsock;
+            vs->lsock = -1;
+            vnc_connect(vs);
+            return 0;
+        }
     }
 
     if (bind(vs->lsock, addr, addrlen) == -1) {
